+++ src/Contracts/Requests/CreateProvince.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Konekt\AppShell\Contracts\Requests;
+
+use Konekt\Concord\Contracts\BaseRequest;
+
+interface CreateProvince extends BaseRequest
+{
+}