+++ src/resources/routes/acl.php
@@ -38,3 +38,11 @@ Route::put('settings', [
     'uses' => 'SettingsController@update',
     'as' => 'settings.update'
 ]);
+
+Route::resource('country', 'CountryController');
+Route::get('/country/{country}/province/create', 'ProvinceController@create')->name('province.create');
+Route::post('/country/{country}/province', 'ProvinceController@store')->name('province.store');
+Route::get('/country/{country}/province/{province}', 'ProvinceController@show')->name('province.show');
+Route::get('/country/{country}/province/{province}/edit', 'ProvinceController@edit')->name('province.edit');
+Route::put('/country/{country}/province/{province}', 'ProvinceController@update')->name('province.update');
+Route::delete('/country/{country}/province/{province}', 'ProvinceController@destroy')->name('province.destroy');